nixos-branding-guide: flip content page every other page
Adds logic to contentPage so that black background content section flips
sides based on page number modulo 2. Intent is to have the black area on
the edge of the page away from the binding when printed. Assumes that
the second page of the guide will be printed on the left.